refactor(vite-plugin-angular): extract virtual modules into shared plugin
Move the duplicated resolveId/load hooks for virtual style/raw modules,
JIT angular:jit: handling, .html?raw/.css?inline interception, and
Vitest fallback into a shared virtualModulesPlugin. This eliminates
the 3x duplication across angularPlugin, fastCompilePlugin, and
compilationAPIPlugin.

Each compilation plugin retains only its component stylesheet registry
mapping in resolveId/load, since that depends on per-plugin state.
